Move Library classes to Lib package
