+++ src/main/java/com/db/lib/SignalHandler.java
@@ -1,4 +1,4 @@
-package com.db.client;
+package com.db.lib;
 
 public interface SignalHandler {
     void handleSignal(int signal);
